Вынес каждый класс в свою вкладку. Исправил переменую int на double. Результат в терминал ограничил двумя числами после запятой

+++ src/main/java/Lesson_5/Fruit/Apple.java
@@ -1,19 +1,19 @@
 package Lesson_5.Fruit;
 
 public class Apple extends Fruit {
-    private static int totalPrice;
+    private static double totalPrice;
 
-    public Apple(int weight, int price) {
+    public Apple(double weight, double price) {
         super(weight, price);
         totalPrice += price();
     }
 
-    public static int totalApplePrice() {
+    public static double totalApplePrice() {
         return totalPrice;
     }
 
     @Override
-    public int price() {
+    public double price() {
         return getPrice() * getWeight();
     }
